First pass of AES TLV packet encryption for Python
A rudimentary first attempt to handle AES256 encryption for Python
meterpreter across varios platforms, on both Python 2 and 3. This code
makes no assumptions about the level of support for crypto on the
target by bringing our own RSA and AES libraries along in
a base64-encoded zip file.

The zip file is extracted in memory and is used as a back end for
a modue resolver.

It's messy and leaves a rather large mess in the code. However, it works
on Windows, OSX and Linux. I'm open to other ideas, but without assuming
that something like PyCrypto is installed we can't really go with
another option.

This was built off the back of a gig that needed us to have encryption
support to bypass MITM stuff on an OSX machine that had Python 2.7.10
with a very old version of OpenSSL, and no presence of PyCrypto or any
other libraries that may have helped.